refactor: adopt layered project structure
