+++ internal/handler/README.md
@@ -1,0 +1,2 @@
+This package will hold handlers (HTTP, CLI, etc.) that expose use cases.
+Business logic should remain in `internal/usecase`.
