+++ internal/usecase/open_stream_usecase_test.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 	"testing"
 
-	"ikedadada/go-ptor/internal/application/usecase"
 	"ikedadada/go-ptor/internal/domain/entity"
 	"ikedadada/go-ptor/internal/domain/repository"
 	"ikedadada/go-ptor/internal/domain/value_object"
+	"ikedadada/go-ptor/internal/usecase"
 )
 
 type mockCircuitRepoOpen struct {
