+++ internal/usecase/service/circuit_build_service_test.go
@@ -6,10 +6,11 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"errors"
-	"ikedadada/go-ptor/internal/application/service"
+	"testing"
+
 	"ikedadada/go-ptor/internal/domain/entity"
 	"ikedadada/go-ptor/internal/domain/value_object"
-	"testing"
+	"ikedadada/go-ptor/internal/usecase/service"
 )
 
 // --- Mock RelayRepository ---
