+++ internal/usecase/shutdown_circuit_usecase.go
@@ -3,9 +3,9 @@ package usecase
 import (
 	"fmt"
 
-	"ikedadada/go-ptor/internal/application/service"
 	"ikedadada/go-ptor/internal/domain/repository"
 	"ikedadada/go-ptor/internal/domain/value_object"
+	"ikedadada/go-ptor/internal/usecase/service"
 )
 
 type ShutdownCircuitInput struct {
